posts_service/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
requests==2.31.0
drf-nested-routers==0.93.4
gunicorn==21.2.0 